MQT-6050 (1) additional checks in ICE candidate negotiation to avoid crash on null string, (2) upgrade CJSON code to fs1.6 version, (3) lossrate change to short, (4) remove some redundant json updates in conf module

+++ src/include/switch_types.h
@@ -767,7 +767,7 @@ typedef struct {
     int last_active_speaker;
     int last_jitter;
     int last_cumulative_lost;
-    float last_lost_percent;
+    short last_lost_percent;
     float last_mos;
     float last_r;
     float last_variance;
